feature FrequencyHistory mainwindow implementation
This part implements 2 actions for navigating back and forward in
frequency history.
The actions are added to tool menu and menubar.
Shortcuts are Ctrl+Shift+Left/Right.

The height of menubar changes from 19 to 24. This is not by intention,
but latest version of Qt Creator sets this by itself and a manual change
back is not possible. There is no visible difference of the window here.