missing drawOverlay() on bookmark marker click
With one of the latest changes there were introduced some
checks to drawOverlay only on changed data.
It turned out that the bookmarker here does count on someone else
draws the overlay, but not anylonger, because no changed
data. Now it draws in right place.

Renamed m_BookmarkTags to m_BookmarkMarker: has nothing to do with
tags.

Using for-range loops.